# WinterwoodStockApi/bin/Release/netcoreapp3.1/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "Winterwood Stock API",
    "description": "Stock/Batch update details",
    "contact": {
      "name": "Kaan Torun",
      "url": "https://www.digitedge.co.uk",
      "email": "KaanTorun@digitedge.co.uk"
    },
    "version": "v1"
  },
  "host": "localhost:44336",
  "schemes": [
    "https"
  ],
  "paths": {
    "/WinterwoodStock/Auth/authenticate": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_RequestToken",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "authModel",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthModel"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/WinterwoodStock/Auth/registerauth": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Register",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "authModel",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthModel"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AuthModel": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
  }
}